chore: refactored into /applicant2 folder
