+++ src/main/steps/applicant2Sequence.ts
@@ -1,17 +1,13 @@
 import { Step } from './applicant1Sequence';
-import { CHECK_ANSWERS_URL, HAS_RELATIONSHIP_BROKEN_URL, HOME_URL, YOU_NEED_TO_REVIEW_YOUR_APPLICATION } from './urls';
+import { HAS_RELATIONSHIP_BROKEN_APPLICANT2, HOME_URL, YOU_NEED_TO_REVIEW_YOUR_APPLICATION } from './urls';
 
 export const applicant2Sequence: Step[] = [
   {
     url: YOU_NEED_TO_REVIEW_YOUR_APPLICATION,
-    getNextStep: () => `${HAS_RELATIONSHIP_BROKEN_URL}2`,
+    getNextStep: () => HAS_RELATIONSHIP_BROKEN_APPLICANT2,
   },
   {
-    url: `${HAS_RELATIONSHIP_BROKEN_URL}2`,
-    getNextStep: () => CHECK_ANSWERS_URL,
-  },
-  {
-    url: CHECK_ANSWERS_URL,
+    url: HAS_RELATIONSHIP_BROKEN_APPLICANT2,
     getNextStep: () => HOME_URL,
   },
 ];
