+++ src/main/steps/your-details/template.njk
@@ -1,3 +1,3 @@
-{% extends "../common/template.njk" %}
+{% extends "common/template.njk" %}
 
 {% block backLink %}{% endblock %}
